updating auth middleware
