+++ client/pages/_app.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import isEmpty from "lodash/isEmpty";
 import { UIShell } from "../components/ui-shell";
 import AuthMegaProvider, { AuthConsumer } from "../components/auth";
-import { authMiddleware } from "../utils/auth-middleware";
+import { cookies, ICookie } from "../utils/auth-middleware";
 
 interface IAppProps extends AppComponentProps {
   token: string;
@@ -13,7 +13,7 @@ export default class Downwrite extends App<IAppProps> {
   static async getInitialProps({ Component, ctx }: NextAppContext) {
     let pageProps = {};
 
-    let token = authMiddleware(ctx);
+    let { DW_TOKEN: token } = cookies<ICookie>(ctx) as ICookie;
 
     if (Component.getInitialProps) {
       pageProps = await Component.getInitialProps(ctx);
