+++ client/utils/responseHandler.ts
@@ -1,4 +1,3 @@
-import Cookies from "universal-cookie";
 import * as Draft from "draft-js";
 import { __IS_BROWSER__ } from "./dev";
 
@@ -6,14 +5,4 @@ export const superConverter: Function = (content: Draft.RawDraftContentState) =>
   return content.hasOwnProperty("entityMap")
     ? Draft.convertFromRaw(content)
     : Draft.convertFromRaw({ blocks: content.blocks, entityMap: {} });
-};
-
-export const getToken = (req, query) => {
-  const ck = new Cookies();
-
-  const token = req
-    ? req.universalCookies.cookies.DW_TOKEN
-    : query.token || ck.get("DW_TOKEN");
-
-  return { token };
 };
